removed sanity check module/moved into airflow
