+++ .github/workflows/runPytest.yml
@@ -25,7 +25,7 @@ jobs:
         python -m pip install --upgrade pip setuptools
         pip install pytest
         pip install -r requirements/tests.txt
-        pip install python/ingestion python/data_server python/datasources python/sanity_checks
+        pip install python/ingestion python/data_server python/datasources
 
     - name: Test python packages with pytest
       working-directory: ./python
