+++ .gitignore
@@ -35,7 +35,6 @@ config/.terraform/
 python/data_server/build/
 python/datasources/build/
 python/ingestion/build/
-python/sanity_checks/build/
 
 /scripts/.env.secrets
 python/tests/datasources/_test_cawp_time_local.py
