--- python/sanity_checks/setup.py
@@ -1,6 +1,0 @@
-from setuptools import setup
-
-setup(name='sanity_checks',
-      package_dir={'sanity_checks': ''},
-      packages=['sanity_checks'],
-      )
